DTSAM-531 move IAC system role tests to IacSystemRoleTest.java
